mc_scraper/pyproject.toml
[project]
name = "mc-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elasticsearch[async]>=9.2.1",
    "meshcore>=2.2.5",
    "pydantic>=2.12.5",
    "pycryptodome>=3.20.0",
]
